Change page/next argument type from Float to Int.

+++ src/main/kotlin/burrows/apps/example/gif/presentation/main/MainPresenter.kt
@@ -36,7 +36,7 @@ class MainPresenter(val view: IMainView,
   /**
    * Load gif trending images.
    */
-  override fun loadTrendingImages(next: Float) {
+  override fun loadTrendingImages(next: Int) {
     loadImages(repository.getTrendingResults(RiffsyApiClient.DEFAULT_LIMIT_COUNT, next))
   }
 
@@ -45,7 +45,7 @@ class MainPresenter(val view: IMainView,
    *
    * @param searchString User input.
    */
-  override fun loadSearchImages(searchString: String, next: Float) {
+  override fun loadSearchImages(searchString: String, next: Int) {
     loadImages(repository.getSearchResults(searchString, RiffsyApiClient.DEFAULT_LIMIT_COUNT, next))
   }
